refactor(codegen): rename system module to call
